Fixed a bunch of bugs uncovered by the RayTracer:
canVectorize now works using a WhiteList. Unsupported operations use a fallback.
Calls are now only copied once if they do not need to be vectorized.
Calls use a fallback if they should, but cannot be vectorized.
fallbackVectorize creates a vectorized version of an instruction using a fallback method. Used for unsupported operations